add replicate-blobs gear (lib + cli)
For publishing to certain chinese AWS-Regions, replication to dedicated
AWS-Buckets is required. Therefore, add replication code intended to be
run as first execution phase prior to executing actual publishing code.

Replication uses identical "keys" (local names within s3 bucket),
and is limited to artefacts required for VM Image Publishing; as it is
as of now only required for AWS, it is only done for a configured subset
of hyperscalers (configuration being set to just AWS).