enhance upload to OpenStack China regions
With https://github.com/gardenlinux/gardenlinux/pull/2134,
Garden Linux build artefacts will be uploaded to an S3 bucket
in AWS China as well, the replication step in glci is thus no
longer necessary and is thus reduced to a check if all required
release artefacts are indeed available in S3 China.

It is now possible to specify individual OpenStack regions
for which release artefacts should not be taken from the main
buildresult bucket but from a replication buildresult bucket
(i.e. in AWS China).