support specify Az gallery regions in publishing_cfg
It is now possible to specify which regions a community
gallery image in Azure should be published in. By
default - if left empty - all regions that are returned
by Azures subscription client will be considered but
it is possible to specify them explicitly.

This feature is necessary for the Gardener integration
test pipeline.